Pass -d to child process
When you set non-default dimensions, the window was being sized
correctly, but the environment variables and such were still
the defaults -- due to using vfork(), the child doesn't get
the modified terminal sizes from the parent.

To fix this, we now explicitly pass the dimensions on to the
child via -d.