Replace all Flips with UpdateRects
The flips were probably being translated into full-screen UpdateRects
anyway, which means we were probably doing a serious amount of
unnecessary pixel copying.

Notably, when flashing the cursor, we should now only be updating
the old and new cursor cells, and when updating the current line,
we should only be updating the current line.